requirements.txt
pyautogui
speechrecognition
pynput
pyaudio
keyboard
pywebview
eel
windows-toasts